requirements.txt
IMAPClient==2.3.1
pyairtable==2.3.3
python-dotenv==1.0.1
Pillow==10.4.0
requests==2.32.3
openai>=1.30.0
pdf2image>=1.17.0
